validation should be a regular function not a hook
