+++ package.json
@@ -156,14 +156,12 @@
     "@babel/preset-typescript": "^7.9.0",
     "@testing-library/jest-dom": "^5.5.0",
     "@testing-library/react": "^10.0.3",
-    "@testing-library/react-hooks": "^3.2.1",
     "@types/jest": "^25.2.1",
     "@types/react-tabs": "^2.3.1",
     "babel-jest": "^25.4.0",
     "jest": "^25.4.0",
     "jest-date-mock": "^1.0.8",
     "jest-styled-components": "^7.0.2",
-    "react-test-renderer": "^16.13.1",
     "ts-jest": "^25.4.0"
   }
 }
